Add a production server for running for reals.
Serves out static files over HTTP in the simplest way possible.

Also added ./start_prod.sh for starting the prod server.
