+++ start.sh
@@ -9,5 +9,5 @@ fi
 
 (cd client && wasm-pack build && cd ../www && npm run start) &
 cargo install cargo-watch
-(cd server && cargo watch -x run -w src/ -w Cargo.toml -w ../core/src/ -w ../core/Cargo.toml) &
+(cd server && cargo watch -x 'run --bin=dev' -w src/ -w Cargo.toml -w ../core/src/ -w ../core/Cargo.toml) &
 (cd client && cargo watch -s 'wasm-pack build' -w src/ -w Cargo.toml -w ../core/src/ -w ../core/Cargo.toml)
