+++ start_prod.sh
@@ -1,0 +1,13 @@
+# When this shell script exits, kill all child jobs.
+trap 'echo $(jobs -p)' EXIT
+
+if test -d "www/node_modules"; then
+  echo 'skipping npm install'
+else
+  (cd www && npm ci)
+fi
+
+echo "Building client..."
+(cd client && wasm-pack build && cd ../www && npm run build)
+echo "Client built, building and starting server..."
+(cd server && cargo run --bin prod)
