+++ www/index.html
@@ -1,11 +1,15 @@
 <!DOCTYPE html>
 <html>
-  <head>
-    <meta charset="utf-8">
-    <title>Hello wasm-pack!</title>
-  </head>
-  <body>
-    <noscript>This page contains webassembly and javascript content, please enable javascript in your browser.</noscript>
-    <script src="./bootstrap.js"></script>
-  </body>
+
+<head>
+  <meta charset="utf-8">
+  <title>rust-us</title>
+</head>
+
+<body>
+  <noscript>This page contains webassembly and javascript content, please
+    enable javascript in your browser.</noscript>
+  <script src="./bootstrap.js"></script>
+</body>
+
 </html>
