+++ www/webpack.config.js
@@ -7,7 +7,7 @@ module.exports = {
     path: path.resolve(__dirname, "dist"),
     filename: "bootstrap.js",
   },
-  mode: "development",
+  mode: "production",
   plugins: [
     new CopyWebpackPlugin(['index.html'])
   ],
